requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
requests>=2.28.0
openpyxl>=3.1.0
selenium>=4.15.0
Pillow>=10.0.0
